SLING-12630: Resource Resolver: refactor MapEntries - extract MapIterator into separate class
